fix: 사이드바 연결
